Split Residents collection resources
